Introduce basic guard conditions as part of `Contract`
Initially, the `Contract` class was modeled after the Code Contracts interface, and sought to maintain interface parity, if not feature parity. Over time, it has evolved to better suit this project's specific needs. As part of that, this update introduces `Requires(object, string)` and `Assumes(object, string)`. These are identical to `Requires()` and `Assumes()` except a) they test the object for whether or not it's `null`, instead of accepting the boolean result of a validity expression, b) they are hardcoded to throw, respectively, an `ArgumentNullException` or an `InvalidOperationException`, and c) the `requiredObject` parameter is flagged with the `[NotNull]` attribute (also introduced as part of this commit) which tells Roslyn's flow analysis know that the input object is guaranteed not to be null—even if it is submitted as such.