Excluded `CA1062` and `CS8602` in `Ignia.Topics.Tests`
The `CA1062` and `CS8602` warnings are useful for identifying potential null values in parameters or externally sourced calls. Adding guard conditions to unit tests isn't strictly necessary, however, since a `NullReferenceException` is a desirable outcome in the case of an unexpected null reference. As such, I'm supressing these errors for the test project.