Reorganized to provide better `null` checking
Explicitly check for `null` `navigationRootTopic` _before_ setting it, as well as failures to resolve the `defaultRoot`. In addition, explicitly ensure that the `navigationRootTopic` is never set to `null` as part of a `while` statement—a condition that will never actually occur at this point, but the compiler doesn't know that.