Workaround for bug (?) in Roslyn's flow analysis
Roslyn's flow analysis doesn't accept the `[NotNull]` hint for parameters unless the variable has been locally assigned, which is an issue for `Requires()` since it is intended to be used exclusively as a guard clause for parameters. Assigning the parameter to itself mitigates this issue—though it does prompt its own warning in return.