Clarified `null` returns with attributes
The `[return: NotNull]` and `[return: NotNullIfNull()]` attributes allow the `GetValue()` methods to be marked as not-nullable or not-nullable based on certain input conditions. So, for example, if a `defaultValue` is provided, then the output will never be `null`, thus providing a useful hint to Roslyn's flow analysis for nullable annotation contexts.