Updated missing `Requires()` and `Assume()` conversions
A previous commit (#3cc15f9) migrated the use of `Requires()` and `Assume()` as guard clauses to use the new overload which provides a `[NotNull]` attribute hint to Roslyn's flow analysis that a given variable will not be null after the execution of the method. A few instances of this were missed due to multiline formatting not picked up by the regular expression or differences in which exception were being thrown (e.g., `ArgumentException` instead of `ArgumentNullException`). This commit adds those to the branch.