Handled `CS8602` warnings ("Dereference of a possibly null reference")
The `CS8602` error occurs when a member of a potentially null object is being called. It can be fixed in a variety of ways: setting defaults, adding guard conditions, adding conditions to avoid the scenario, or providing compiler hints via attributes so that Rosyln's flow analysis understands the conditions under which a null will not occur. This commit employees all four strategies to eradicate all remaining `CS8602` errors.