Excluded `CS8604` from test project
The `CS8604` concerns passing a potentially `null` value to a method call that doesn't expect `null` values when nullable annotation contexts are enabled. Because the project is currently targeting .NET Standard 2.x and .NET Framework 4.8, the underlying class libraries don't include the appropriate nullability annotations yet. This generally isn't a problem, or can be handled on a case-by-case basis, but it does introduce issues with the test project since most method calls on the `Assert` class may be passed `null` values as part of testing. As such, these warnings are suppressed until the project is upgraded.