Annotate nullability on models
Currently, the `ITopicMappingService` doesn't recognize constructors, and instead populates properties by direct assignment. This means that properties must default to being `null`. In addition, the `TopicMappingService` implementation tends to take a forgiving approach, simply not mapping properties that it can't resolve instead of e.g. throwing an exception. Worst case, the views may be missing content. Given this, the vast majority of properties on view models have been set to nullable. That said, if a value is required, they can still use the e.g. `[Required]` attribute, which will check for a non-empty value; in that case, the property may be null, but an exception can still be raised. In the future, we may revisit constructor support, in which case properties we expect never to be `null` (e.g., properties that are required by the content type) may be reevaluated. This addresses many of the `CS8618` warnings related to non-nullable properties being uninitialized.