Introduced missing guard clauses
Identified a few cases where guard clauses were either missing, or where they were incomplete (e.g., checking for nullability of a parameter, but not properties on the property). In most of these cases, nulls are not expected in real-world conditions (e.g., as the result of a SQL reader from a sproc), but having them a) quiets warnings from Roslyn's flow analysis within a nullable annotation context, and b) ensures that if the unexpected occurs, there's a somewhat more useful error handling.