Migrate to *.NET Standard 2.1* and *ASP.NET Core 3.0*
With the release of *ASP.NET 3.0* today, we're able to (officially) migrate the `Ignia.Topics.AspNetCore.Mvc` project to `netcoreapp3.0`, and the class libraries up to `netstandard2.1` (which ships with *ASP.NET 3.0*). We were already using *C# 8.0*, but this allows us to use the full API of *.NET Core 3.0*, and also (finally!) build our nullable annotation context in a debug release without enabling previews.