Updated `MSTest` to 2.0
The `MSTest` framework has been updated to the next major release—2.0. As a result, both the `MSTest.TestAdapter` and the `MSTest.TestFramework` versions have been updated from 1.4.0 to 2.0.0. This not only includes significant performance updates, but also accounts for a *C# 8*'s nullability annotation context, thus eliminating a number of warnings (~25 issues). This also includes a subsequent update to `Microsoft.NET.Test.Sdk` to 16.3.0, which updates the targets and project defaults to correspond with the release of *Visual Studio 16.3.0*.