Suppress `CA1303` warnings
The `CA1303` warning pertains to string literals in `Exception` messages, which it advises moving into a resources file. This is probably a good idea in the long-term. Since a) most of these strings are unique error messages today, and b) there aren't any current plans to localize the *Ignia OnTopic Library*, these are just adding clutter. As such, they're being suppressed. In addition, `1701` and `1702` were listed in the properties windows of most projects, but not in the `CSProj` file; it's unclear why this is, but saving the updates persisted these to file. Presumably, these were implicit defauls previously.