Added null-conditional operators
In several cases, *C# 8*'s nullability context discovered potential issues dereferencing a null object which could easily be avoided by simply using *C#*'s null-conditional operators (e.g., `?.`). In this case, a `null` is the result instead of a` NullReferenceException`. This only works in particular scenarios where a null value is acceptable, or there is an acceptable default value.