Work around for `IsNullOrEmpty` escape conditions
For reasons I'm unclear on, *Roslyn*'s flow analysis doesn't (always) recognize escape conditions using `IsNullOrEmpty()` as guaranteeing that a variable is subsequently null. These are mitigated by placing an explicit null check in the condition. This is a bit ugly since this is repetitive with the `IsNullOrEmpty()`; we may want to reevaluate this at a later date.