Flag `GetTopic()` as never null
Generally, we should avoid flagging items as never null (`!`), as it often hides a potential issue. In this case, however, the issue is because we're using `Contains()` for the guard clause, but `GetTopic()` for the actual call. This makes sense as `GetTopic()` will throw an exception if the `Key` is invalid, which is irrelevant if the `Key` doesn't even exist. (Though, admittedly, this isn't a scenario we'd normally _expect_ to happen.)