Refactor conditional logic as an escape clause
The conditional logic is a bit cumbersome to read, and it's not immediately clear what each clause is for. To mitigate this, refactored this into an escape clause with a comment explaining the logic. There was also a scenario that _should_ be fine, and _is_ handled by the `SetParent()` logic. It's unclear why this was included. This is a _potentially_ breaking change that will require explicit integration testing.