Replaced `IsNullOrWhiteSpace()`, `IsNullOrEmpty()` instances
Commit #ca9eb06 established a workaround for the fact that Roslyn's flow analysis fails to recognize `IsNullOrEmpty()` (or `IsNullOrWhiteSpace()` escape clauses as preventing null dereferences. The initial fix was ugly: It redundantly checked for null and _then_ maintained the `IsNullOr*()` check. This commit addresses that by replacing the `IsNullOr*()` check with `Length.Equals(0)` or `Length.Trim().Equals(0)`. I'd _prefer_ to use the shortcuts, but it reads as sloppy when they're redundant with the previous logic. Hopefully the underlying issue will be resolved in the future.