Ensured disposable objects call `Dispose()`
A number of disposable objects—and, in particular, controllers in the `TopicControllerTest` class—were not being properly disposed of. This was especially a concern in the `SqlTopicRepository` where a `SqlCommand` object that is iterated through in the editor was not being closed properly, as that could result in memory leaks by deferring garbage collection.