Removed unused `ConvDbType()` private function
Previously, `ConvDbType` was used to convert the string names of SQL types to their `SqlDbType` enum value. This was a poor approach for a number of reasons: a) it relied on a string value that couldn't be verified at compile time, and b) the switch statement could easily be replaced with a single enum match, if necessary. That said, all calls to this have long been replaced with use of the enum itself, so this is no longer used or necessary. Good riddance!