Suppressed `IDE0059`
The `IDE0059` error pertains to variables that are initialized but never used, or initialized but redefined before they're ever called. It's a good warning. But there are a lot of Unit Tests where this is done to maintain a consistent structure and make it easy to reliably call variables should they be needed. That's perhaps of questionable judgment, but it's a trivial cost and not currently a priority to fix. (In most cases, the value is assigned by a function that is under test; as such, the call was necessary independent of whether it needed to be assigned to a variable. If this wasn't the case, the variable was separately removed as part of commit #1ffaa88).