Removed unused `ITypeLookupService` dependency
I mistakenly introduced an unnecessary dependency on `ITypeLookupService` in the `ReverseTopicMappingService`'s constructor. The class _does_ do type lookups using reflection, but _not_ within the type of constrained scope that really benefits from the functionality of `ITypeLookupService`. E.g., there's usually not going to be a fixed or deterministic list of source types known to the configuration root. We may revisit this later, but since it's not used today it shouldn't be maintained as an unnecessary placeholder. (As a bonus, no clients have yet implemented the `IReverseTopicMappingService`, since it hasn't made it to an official release, so breaking the public signature won't have a downstream effect.)