Validate `ITopicRepository` results in test
The `TopicControllerTest` constructor called `CachedTopicRepository.Load()` and assumed a not-null result. In practice, this should be fine since it's working off of a stub. However, *Roslyn*'s flow analysis doesn't know this, so we're explicitly declaring that this is known to return a result. (In a normal dependency, this would not make sense; this only works since we're explicitly working off of a concrete instance of a fake.)