Suppressed `CA2227`
Generally, collections should be explicitly declared and not include a setter. This is what `CA2227` warns us about. This is because, generally, customers expect a collection to be defined, and won't need to replace it; instead, they might clear it, or replace all items in it. That isn't the case for the `CompatiblePropertyTopicViewModel`, however, as it's explicitly intended to have the value set by the `ITopicMappingService`, assuming a matched signature.