Suppressed `CS1717` warning
The `CS1717` warning flags instances of variables being set to themselves. That's a good thing generally. In this case, however, it's a very deliberate `###HACK` around a(n apparent flow in) *Roslyn*'s flow analysis. Without this, it fails to recognize that the return value is not actually null—even though that's effectively what the next line does. Since this is a known and well-documented `###HACK`, I've suppressed the warning.