Suppressed `CA1307` error
In *.NET Core*, there is an overload of `Replace()` that accepts a `StringComparison` enum to specify whether or not it should be case sensitive, and whether or not it should honor the current culture. `CA1307` suggests using this to avoid unpredictable behavior with globalization. That said, this overload isn't available in *.NET Framework* which implicitly uses `InvariantCultureIgnoreCase`. As a result, this also isn't available in *.NET Standard*. Unfortunately, *Roslyn*'s code analysis isn't aware of this disconnect when *.NET Standard*. This is a known bug (see dotnet/roslyn-analyzers#1552). Until this issue is resolved, we're explicitly supressing these warnings.