Fix instantiation of `ITopicRepository` dependency
A previous commit, #4ac8b9b, modified the call to `ITopicRepository.Load()` to use the locally defined `_uri` variable, instead of hard-coding the target topic path and, thus, effectively repeating the `_uri`. That's fine, except that this necessitated reordering the variables, which hadn't been done. This caused a runtime error when running the test, which this commit fixes.