Introduced `[ValidatedNotNull]` attribute
The `[NotNull]` attribute communicates to *Roslyn*'s flow analysis that a reference parameter should not be `null` after the method completes, which is useful for e.g. validation methods.

That is enough to satisfy *Visual Studio*'s OOTB validation of *C# 8*'s nullable annotation contexts. Unfortunately, however, this doesn't satisfy *Code Analysis*' `CA1062`, which only recognizes `[NotNull]` when it's in the _local_ assembly. For ancillary assembly's it requires `[ValidatedNotNull]`—as discussed in e.g. dotnet/roslyn-analyzers#2578.

To mitigate this, a dependency on `Microsoft.VisualStudio.Validation` is introduced—though a custom attribute of the same name would also do. In addition, public methods using `[NotNull]` were annotated with `[ValidatedNotNull]`.

Note that `CS8604` still only recognizes the `[NotNull]` attribute. As such, it is still needed for scenarios where an otherwise potentially-nullable reference is being passed to a method that does not accept nullable parameters. Since this is not a predictable use case for a public library, that effectively means that all instances of `[ValidatedNotNull]` must _also_ include `[NotNull]`.