Removed self-assignment hack in `Requires()`
To work around an apparent issue in *Roslyn*'s flow analysis, the `requiredObject` parameter of the `Requires()` method was set to itself in #edf39c2. This mitigated a number of warnings, but introduced its own `CS1717`, which was suppressed in #64e2a06. That issue has been resolved by introducing the `[ValidatedNotNull]` attribute, as per #b9e0151. Given that, the original hack—as well as the suppression of its subsequent warning—can be removed. Good riddance!