Remove unused `PersistRelations()` overload
While the `PersistRelations(Topic, SqlConnection)` overload could potentially provide some value, it isn't currently being used, it isn't difficult to add the third `bool` argument, and, as a private helper method, there's no use keeping it around in case we potentially need it in the future. Besides, if we do, it'd be easier to simply assign a default value to the `skipBlob` parameter in the remaining overload.