Made `sourceTopic` nullable for `IHierarchicalTopicMappingService<T>`
The code for the `IHiearchicalTopicMappingService<T>` is written to account for the `sourceTopic` to be `null`. But when I implemented **C#**'s nullability contexts, I marked it as required. This was a mistake. As with the `ITopicMappingService`, there will be times where a topic is not known—and instead of putting guards before every call, it makes more sense to just put guards in the service and return a null object. In fact, this code already exists, so it just needed the `soureTopic` parameter updated to use `Topic?` throughout the interface.