Removed requirement for `navigationRootTopic`
In commit #0a19bb1, I made it possible to pass a null `sourceTopic` to the `IHierarchicalTopicMappingService<T>` implementations. This was especially in service of `PageLevelNavigationViewComponentBase<T>`. A concrete instance of this is typically included on every page via a `_layout.cshtml`. But not every page is of the `PageGroup` content type used for page-level navigation. With the `IHierarchicalTopicMappingService<T>` configured to allow null `sourceTopic`s, we can now remove the `Contract.Assume()` around `NavigationRootTopic` in the `PageLeelNavigationViewComponentBase<T>`, thus allowing `null` values to simply flow through and be (easily) accounted for in the view, as oposed to throwing an exception.