Removed `AddControllersAsServices()`
In **ASP.NET Core 2**, `AddControllersAsSerices()` was necessary for the controllers from an external assembly to be recognized. This is no longer true in **ASP.NET Core 3**. Further, and perhaps more importantly, this introduces _problems_ with **ASP.NET Core 3** because it will, by default, attempt to validate the dependency structure of all services. Since customers are not necessarily configuring the **ASP.NET Core** dependency injection container, this causes problems. As such, this is being removed. If a customer _is_ using **ASP.NET Core**'s DI container, they should add this back in via their own `Startup` class.