Added `MapTopicRoute()` for `IEndpointRouteBuilder`
There is already a `MapTopicRoute()`, but it targets `IRouteBuilder`. In **ASP.NET Core 3**, endpoint routing is preferred over `IRouteBuilder`. As such, the new extension method is functionally identical to the previous one, but supports `IEndpointRouteBuilder`. The legacy version is maintained for customers that either can't migrate to endpoint routing yet, or who are still using **ASP.NET Core 2**.