Disable bypass of hidden topics
We _should_ have the _ability_ to skip hidden topics—this feature, introduced in #885d9a6d, makes perfect sense. However, current implementations rely on this as a way of displaying an index of otherwise-hidden topics. As such, this needs to be made conditional. Until we have a better solution (such as a `[IncludeHidden]` attribute?) this should be excluded.