Prevent `IsDisabled` or `List` mappings
Topics that were marked as `IsDisabled` were already disallowed, but this only happened later, after an empty object had been created. Topics of the content type `List` are skipped as part of `PopulateChildTopic()`, but there are plenty of backdoors exposed by refactoring of the new(ish) `SetCollectionValueAsync()` logic. To avoid either of these scenarios inadvertently happening, an escape condition is introduced upfront. Since this could also result in an null object being added to collections, this condition is also handled. (In the past, an empty object would have been added, which isn't any more desirable.)