Introduced support for `NavigationRoot`
The `MenuViewComponentBase<T>`, by default, will crawl up the directory from the `CurrentTopic` until it hits a root namespace (e.g., `Web`) and then display children underneath that root namespace in the menu. Sometimes, however, you may want to display menu items from a _different_ namespace. For example, if you have a separate namespace (e.g., `Forms` or `Library`) then you may want to display the contents from `Web` instead. To accommodate this, `MenuViewComponentBase<T>` will crawl up the topic tree and look for a `NavigationRoot` attribute—which is now defined as an attribute on the `Container` content type descriptor. If found, that will be used as the navigation root. Otherwise, it will fall back to the original behavior.