Fixed bug with truncating `@Attributes`
While the `Split` function is potentially used in many areas, one of its primary purposes in the **Topic Library** is parsing a delimited list of attribute key/value pairs in `topics_CreateTopic`. Unfortunately, however, while the sproc accepts a `varchar(max)` for the `@Attributes` parameter, the `Split` function limits the string to 512 characters. In practical terms, this has generally been acceptable since we usually only have a handful of "indexed" attributes (i.e., those not stored in the `topics_Blob` table). However, as we added more indexed fields to some customer's content types, this caused truncation of data—and, in some cases, could even result in database corruption (e.g., by creating an attribute with a blank `Key`, depending on where the content was truncated). For some reason, the `with` operation fails with `varchar(max)`, which requires further investigation. Setting the function to accept the maximum size of `8000` works, however. Even with `varchar(8000)` there's a theoretical maximum of 15 indexed fields all using the maximum 256 characters for both fields and labels. In practice, however, we don't expect most customers to hit this limit; not only is it rare to have that many indexed fields, but the values are rarely the maximum allowed length. That said, we may explore migrating this to a `text` field if we start expanding the length and use of indexed attributes in the future.