Introduced custom collection mapping
Introduced new functionality to `TopicMappingService` such that if a target collection maps to a source collection, and the `RelationshipType` either isn't set, or is set to `Any`, then it will attempt to convert the (presumably) `Topic` instances in that collection to `TopicViewModel`s and, assuming they are compatible with the target type, add them to the target collection. This is useful for custom collections on derivatives of `Topic`, which may not map to normal relationships, such as `ContentTypeDescriptor.AttributeDescriptors`.