Introduced basic type validation
It's a bit cumbersome to validate the collection type if it's not known ahead of time if the collection is generic, or not. And while all collections we'll be working with _are_ generic under the hood, the actual derived class may not be. Given that, we're at least providing basic validation by evaluating the type of the first item in the collection before continuing with mapping. This could still fail in the case that a collection is mapped to e.g. `Object`, and the _first_ entry _happens_ to be a `Topic`, but others are not. That's an unlikely scenario, however—and not one we'd expect a developer to introduce. Still, we may improve this error handling in the future.