Improved mapping (and testing) of compatible properties
Previously, the compatible property testing would fail for properties that implemented `IList` as they would be caught be a separate condition. And there is no efficient way of validating after the fact whether or not a value has been set. As such, we rearranged the logic so that `SetCompatibleProperty()` is tried up front, and returns `true` if it succeeds. If so, no further processing is required. This allows 1:1 mapping to easily be tested upfront before doing any custom mapping. In addition, strengthened the unit tests to evaluate the `IList` scenaril, since it wasn't being factored in previously.