Introduced attribute type descriptors
Previously, attribute types were handled by a `FileList` attribute on the `AttributeDescriptor` content type. That prevented the ability to add attribute type specific attributes to a given `AttributeDescriptor`. To mitigate that, we're introducing a new content type for each attribute type, which will derive from `AttributeDescriptor`. These don't actually contain any values; the only purpose of them is to ensure that topics created as e.g. a `BooleanAttribute` derive from `AttributeDescriptor` and, thus, have access to _its_ members. These are _also_ added to the `DefaultTopicLookupService`, which is generally registered with the `TopicFactory`, so that attributes will be correctly creating using their derived content types when e.g., `ITopicRepository.Load()` calls `TopicFactory.Create()`.