Added support for (explicitly) mapping `List` content types
Previously, the `TopicMappingService` prevented any `List` topics from ever being mapped, as per #fe1f117. In practice, however, there are corner cases where this may be needed, such as referencing topic containers via topic references—a scenario that occurs in the editor, for instance. `List`s are still ignored if they are in a child container, but otherwise they are processed.