Introduced tests for evaluating nullable types
Nullable types should be mappable to non-nullable types. If they are mapped to an attribute which is null, empty, white space, or cannot be converted, then they should retain their existing value (as opposed to being overwritten with a presumed default). The first test passes, as these properties aren't currently being mapped (a bug). The second test fails for the same reason.