Introduced support for (un)following `MappedCollections`
While support for mapped collections was previously introduced, it has always mapped those collections—potentially recursively—without limit. This is in contrast to other types of collections which require an explicit `[Follow()`]` attribute to map those collections more than one level deep. Likewise, it didn't permit annotations to help restrict the type of source it should be mapped to. Both of these are fixed with the introduction of `Relationships.MappedCollections`, `RelationshipType.MappedCollection`, a mapping between them in `RelationshipMap`, and an explicit check for these in `TopicMappingService` (already built into the helper function `GetRelationship()`, but not previously wired up for this collection type).