Introduced new `[MapToParent]` attribute
The new `[MapToParent]` attribute allows a complex object type to be assigned to a property on a binding model, with its properties being mapped to the attributes (or relationships) of the target topic. This allows for the creation of more sophisticated binding models which may facilitate reusable code, while continue to support mapping back to a single topic. The `[MapToParent]` attribute optionally accepts an `AttributePrefix` property which specifies a prefix that should be used when mapping properties on the complex object back to topic attributes; by default, this will be the name of the property that the complex object is set to, but it can optionally be a different name—including an empty string.