Fixed issue with `TopicLookup` being treated as a `Reference` type
Technically, a `TopicLookup` (now, `TopicList`) _could_ be a reference type—assuming it's saving either the `{TopicId}` or `{UniqueKey}`—but this isn't guaranteed, and often isn't the case. If we want to treat this as a reference type, this is going to require additional work. In fact, really, the entire concept of mapping references probably needs more work as, often, these may in fact just be integers or strings, as opposed to separate objects, which should be accounted for. But, for now, this at least prevents the very common scenario of `TopicLookup` lists being misclassified as a `ModelType.Reference`. (The setup of `ModelType` and `EditorType` _also_ needs additional reconsideration—but also a task for another day.)