Ensured topic references honor `[AttributeKey()]` attribute
Previously, topic references were hardcoded to assume the property name was the attribute name without the `Id` suffix. This is a reasonable convention. But there are two problems with it. First, it doesn't honor the `[AttributeKey()]` attribute. Second, it doesn't support cases where the value actually ends in `Id` (as would be a logical assumption when using [AttributeKey()]`. This update resolves both issues. Note that we should ultimately want to establish a `[Relationship()]` type for annotating these, but that's not an immediate priority.