Expanded view locations to prioritize `/Controller/Action` folders
While `/Views/Controller/Action.cshtml` is already part of the default search path, we want to search it before falling back to e.g. `/Views/Shared/View.cshtml` fallbacks. Since the precise order is so important, as to avoid generic fallbacks from intercepting a more specific view, this is being injected into the `TopicViewLocationExpander`. In addition, comments were added to each path so it's easier to quickly read what they're referring to.