Implemented new `ITopicRepository.Load(RouteData)`
With the new `ITopicRepository.Load(RouteData)` extension method (see #15eeebe), we can now get rid of the legacy (and now deprecated) `MvcTopicRoutingService`. This doesn't reduce the dependencies—now, instead of an `ITopicRoutingService`, `TopicController`, its derivatives, and `NavigationTopicViewComponentBase<T>` classes will instead require a direct dependency on `ITopicRepository`. This is, of course, a breaking change, since it updates the dependencies of the public API.