Establish `contenttype` route variable
When we deprecated and replaced `MvcTopicRoutingService` (see #15eeebe and #1d09cfb), we lost the setting of the `ViewData.Values["contentype"]` key which is required in order for the `TopicViewResultExecutor` to function properly, as it doesn't have access to e.g. the `TopicViewResult`, only the `RouteData`. Ideally, we want this to be set somewhere in the infrastructure so it's not dependent on implementors to set. The perfect place for this is the `TopicViewResultExecutor` since we don't expect the expanded views (e.g., `/Views/ContentType/View.cshtml`) to be used or needed except in cases where a `TopicViewResult` was returned from the `Controller` (and, specifically, a `TopicController` or derivative).