Search for view by action name
Typically, with a default implementation of the `TopicController`, the view is contextual to the request, the topic, or the content type. In derived cases, however, the action name may be relevant. For instance, on a `FormController` that derives from `TopicController`, an action `Contact` should first look for `/Views/Form/Contact.cshtml` before looking for e.g. `/Views/FormData/FormData.cshtml` (i.e., based on the content type). This is similar to how **ASP.NET Core** searches for views out-of-the-box.