Introduced barebones `Host` project for testing
Previously, integration testing required packing the topic libraries, publishing them to Ignia's internal **NuGet** repository, updating e.g. the  `Topics-AspNetCore-Sample` site to pull the **NuGet** packages, and testing. If there were bugs, there wasn't an easy way to debug them, and the process needed to be repeated once they were resolved. By introducing a barebones `Host` project, we allow local testing, which is far more efficient. This is also handy because, as opposed to the `Topics-AspNetCore-Sample` repository, which seeks to provide a fully-fleshed out example, this version is as barebones as possible; the views, for instance, just list out properties, but provide no formatting, stylesheets, or javascript. As such, this allows integration testing without distraction, while simultaneously avoiding (too much) duplication of code.