Reduced prior expansion of topic search paths
In a previous commit (#32d1ef5) the scope of the recent `ITopicRepository.Load()` extension method (#1d09cfb) was overly aggressive, allowing a fallback to the `{Controller}` even when a path was specified and non-existent. This is not the desired behavior in most circumstances. If it's desired in specific cases, that needs to be handled through either an overload or some other form of logic. Until such a circumstance is identified, we're reducing this prior expansion to account for the bulk of scenarios.