Delete legacy `FallbackController`
The `FallbackController` doesn't really provide that much of a function. **ASP.NET Core** already has handling for the case where a controller can't be identified. And, even in the case where a controller is found but not supported by the composition root, it's completely appropriate to e.g. throw an exception in that case.