Delete legacy `ErrorControllerBase<T>`
The `ErrorController` creates an unnecessary level of complexity which is better handled by either just using the `TopicController` (e.g., by mapping it to an `/Error` route), or by using static pages (in the case of e.g. a database exception). In practice, this is usually wrapped in the normal `_Layout.cshtml` which is why we need to deliver a faux `IPageTopicViewModel`, but if we're doing that we might as well just map to a _real_ topic and thus allow the customer to modify the error page via the editor, instead of having it be hardcoded in a `cshtml` file. (Acknowledging that for exceptions pertaining to the `ITopicRepository` or `_Layout.cshtml`, we won't be able to use a dynamic page.)