Moved sitemap view logic inline
Previously, the majority of the sitemap logic was in the view, which was simply responsible for generating XML. There were several problems with this. First, it pushed the logic down to the implementation, whereas we'd prefer for it to be centralized. Second, that put a lot of logic in the view, which it isn't well suited for. Third, it was generating static text in the format of XML, which introduced potential issues with e.g. encoding. All three of these are resolved by moving this logic into the `SitemapController` itself and then returning the output of an `XDocument`. There are probably still some further optimizations that could be made to this in terms of how the `XDocument` is constructed, but it's a big improvement over the legacy approach.