Replaced `partition by` with `group by` in `TopicVersions` query
This resolves an issue where the ranking was not over a distinct list, and thus a topic with more than six (indexed) attributes would end up only showing one version. Whoops. By using `group by`, we're able to rank over a distinct list of `TopicId` and `Version`, and then select the top six of those.