Fixed bug in areas search
The `AreaViewLocations` search, which had never actually been used, was incorrect in that it failed to check in the `/Areas` folder, instead assuming the area was in the root. That is inconsistent with **ASP.NET Core** conventions. We may want to explore alternate area search paths in the future, but that wasn't the intent of this update.