Reset SQL to tab width of 16
We generally avoid tabs. One exception is in SQL, where much of our alignment is by columns of 16. Here, we use a tab width of 16, while continuing to use spaces for indenting logical blocks. Ideally, we'd just set the `.editorconfig` to `tab_width=16`. Unfortunately, **Visual Studio** doesn't recognize/honor this. So, instead, the `indent_style` and `indent_size` are specified.